Aestetics Changed
Where do i being. I changed the buttons, and the background color for many screens.. I also had to change the text color. Here is where the issue lies... I found several bugs on the fact that the button is still clickable on the wait for green state... worse it shows soem number on the reaction time... Also, the play section doesnt work either... sooo. first I got to fix the practice section....